onestock/requirements.txt
pytest
flake8
Flask
gunicorn
pyfunctional
google-cloud-firestore
google-cloud-bigquery
pyyaml
pytz
httplib2